+++ src/index.tsx
@@ -6,6 +6,7 @@ export { useConfirmSetupIntent } from './hooks/useConfirmSetupIntent';
 export { useStripe } from './hooks/useStripe';
 export { useApplePay, Props as UseApplePayProps } from './hooks/useApplePay';
 export { usePaymentSheet } from './hooks/usePaymentSheet';
+export { useGooglePay } from './hooks/useGooglePay';
 
 //components
 export {
@@ -14,6 +15,7 @@ export {
   Props as StripeProviderProps,
 } from './components/StripeProvider';
 export { CardField, Props as CardFieldProps } from './components/CardField';
+export { CardForm, Props as CardFormProps } from './components/CardForm';
 export {
   ApplePayButton,
   Props as ApplePayButtonProps,
@@ -26,6 +28,10 @@ export {
   StripeContainer,
   Props as StripeContainerProps,
 } from './components/StripeContainer';
+export {
+  GooglePayButton,
+  Props as GooglePayButtonProps,
+} from './components/GooglePayButton';
 
 export const events = new NativeEventEmitter(NativeModules.StripeSdk);
 export * from './functions';