+++ src/types/index.ts
@@ -6,6 +6,7 @@ import type {
   ConfirmSetupIntentError,
   CreatePaymentMethodError,
   CreateTokenError,
+  GooglePayError,
   PaymentSheetError,
   RetrievePaymentIntentError,
   RetrieveSetupIntentError,
@@ -21,10 +22,12 @@ export * from './ApplePay';
 export * from './PaymentIntents';
 export * from './PaymentMethods';
 export * from './SetupIntent';
+export * from './GooglePay';
 export * from './ThreeDSecure';
 export * from './components/ApplePayButtonComponent';
 export * from './components/AuBECSDebitForm';
 export * from './components/CardFieldInput';
+export * from './components/CardFormView';
 export * from './Card';
 export * from './Errors';
 export * from './PaymentSheet';
@@ -78,7 +81,7 @@ export type RetrieveSetupIntentResult =
       error: StripeError<RetrieveSetupIntentError>;
     };
 
-export type ConfirmPaymentMethodResult =
+export type ConfirmPaymentResult =
   | {
       paymentIntent: PaymentIntent;
       error?: undefined;
@@ -182,3 +185,37 @@ export interface InitStripeParams {
 export interface InitialiseParams extends InitStripeParams {
   appInfo: AppInfo;
 }
+
+export type GooglePayInitResult =
+  | {
+      error?: undefined;
+    }
+  | {
+      error: StripeError<GooglePayError>;
+    };
+
+export type PayWithGooglePayResult =
+  | {
+      error?: undefined;
+    }
+  | {
+      error: StripeError<GooglePayError>;
+    };
+
+export type CreateGooglePayPaymentMethodResult =
+  | {
+      paymentMethod: PaymentMethod;
+      error?: undefined;
+    }
+  | {
+      paymentMethod?: undefined;
+      error: StripeError<GooglePayError>;
+    };
+
+export type OpenApplePaySetupResult =
+  | {
+      error?: undefined;
+    }
+  | {
+      error: StripeError<ApplePayError>;
+    };